Make component more flexible
Use event and array ID on nameChangedHandler where we'll compare the ID of the element. If it's true then we'll copy the array and pass in that index then assign event.target.value to the name and create a new array with those updates. This will continue to update the text that our input field will change.